added rule filter

+++ generate_rules.py
@@ -1,17 +1,27 @@
 import json
 
 file = json.load(open('rules_train.json'))
+good_rules = json.load(open('rules_w_prec_greater_60.json'))
 
 total = 0
 
 for relation in file:
+    if good_rules and relation.replace('_slash_', '/') not in good_rules:
+        # print (file[relation])
+        continue
+    if good_rules:
+        current_rules = sorted(good_rules[relation.replace('_slash_', '/')], key=lambda k: len(good_rules[relation.replace('_slash_', '/')][k]), reverse=True)
+        if len(current_rules)>1:
+            print (relation, current_rules)
     with open('%s_unit.yml'%relation, 'w') as f:
         count = 0
         for trigger in file[relation]:
             rules = [r for r in file[relation][trigger]]
             relation = relation.replace('/', '_slash_')
+    
             for rule in rules:
-                f.write('''
+                if good_rules==None or count in good_rules[relation.replace('_slash_', '/')][current_rules[0]]:
+                    f.write('''
 - name: ${label}_${count}_%d
   label: ${label}
   priority: ${rulepriority}
@@ -19,7 +29,19 @@ for relation in file:
     trigger =  %s
     subject: ${subject_type} = %s
     object: ${object_type} = %s\n'''%(count, trigger, ' '.join(rule['subj']), ' '.join(rule['obj'])))
+                    total += 1
+                if good_rules and len(current_rules)>1:
+                    for c in current_rules[1:]:
+                        if count in good_rules[relation.replace('_slash_', '/')][c]:
+                            f.write('''
+- name: ${label}_%s_%d
+  label: ${label}
+  priority: ${rulepriority}
+  pattern: |
+    trigger =  %s
+    subject: ${subject_type} = %s
+    object: ${object_type} = %s\n'''%(c, count, trigger, ' '.join(rule['subj']), ' '.join(rule['obj'])))
+                            total += 1
                 count += 1
-                total += 1
 
 print (total)
